apub: handle collection recipients mroe reliably
When we receive activity that has a collection, we see if any of the
other actor recipients have the collection set as their followers.
Then we can encode it with plus addressing. If we can't do that we
drop it. Not ideal, but it covers most use cases.

Next up we can encode arbitrary collections like
"some/path/to/thing@example.com".